feat: Add likes and unlikes attributes to Question model
This commit adds the `likes` and `unlikes` attributes to the Question model in order to calculate the total number of likes and unlikes for a question. These attributes are defined as computed attributes using the `Attribute` class from the `Illuminate\Database\Eloquent\Casts` namespace. The `likes` attribute calculates the sum of all the 'like' votes associated with the question, while the `unlikes` attribute calculates the sum of all the 'unlike' votes. This enhancement allows for easy retrieval of the total likes and unlikes for a question.